feat(sla-page): add back navigation to SLA details
Implement a back button on the SLA details page to improve user
navigation. The button's destination is determined by a new 'from'
query parameter, allowing users to return to the Plans or Packages
page from which they navigated. Links from the Plans and Packages
pages have been updated to include this parameter.

Signed-off-by: SkyVPS360 <201673712+skyvps360@users.noreply.github.com>